Worked on parser + lexer till figured out globals were necessary
The tutorial's scan method modifies global token & value variables.  I
thought I could easily rewrite this so that scan simply returns an
object with both values, but late in integrating the lexer with the
parser found code that reaches up to those global variables & gave up on
my design.  I'll have to give in and use the globals in the next
installment